@@ -13,31 +13,25 @@ description: my Flutter mobile project
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 1.0.0+1
 
-author: bob steele
 homepage: https://bsteele.com/index.html
 
 environment:
-  sdk: ">=2.5.0 <3.0.0"
+  sdk: 2.14.0-383.0.dev
 
 dependencies:
   flutter:
     sdk: flutter
-  logger: ^0.8.3
-  quiver: '^2.0.1'
-  #basic_utils: ^2.0.2
+  logger:
+  quiver:
   source_gen:
-  build: ^1.0.0
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  build:
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.0.0
-  build_test: ^0.10.3
-  build_web_compilers: any
+  build_runner:
+  build_test:
+  build_web_compilers:
 
 
 
